automatically generate go sdk docs
