+++ package.json
@@ -16,7 +16,8 @@
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
-    "typecheck": "tsc"
+    "typecheck": "tsc",
+    "update:sdk-docs": "docusaurus download-remote-sdk-content"
   },
   "dependencies": {
     "@docusaurus/core": "^2.4.1",
@@ -33,6 +34,7 @@
     "aos": "^2.3.4",
     "autoprefixer": "^10.4.14",
     "clsx": "1.2.1",
+    "docusaurus-plugin-remote-content": "^3.1.0",
     "docusaurus-plugin-sass": "^0.2.3",
     "docusaurus-theme-github-codeblock": "^1.1.4",
     "instantsearch-itemsjs-adapter": "^1.1.5",
@@ -50,7 +52,9 @@
     "remarkable": "2.0.1",
     "remarkable-react": "1.4.3",
     "sass": "^1.63.6",
-    "tailwindcss": "^3.3.2"
+    "tailwindcss": "^3.3.2",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "2.4.1",
