+++ tsconfig.json
@@ -1,7 +1,16 @@
 {
-  // This file is not used in compilation. It is here just for a nice editor experience.
   "extends": "@tsconfig/docusaurus/tsconfig.json",
   "compilerOptions": {
-    "baseUrl": "."
+    "baseUrl": ".",
+    "resolveJsonModule": true,
+    "esModuleInterop": true,
+    "lib": [
+      "ES2021"
+    ],
+    "paths": {
+      "@site/*": [
+        "./*"
+      ]
+    }
   }
 }
